Fixed some Typos, the build Warnings, NPC Gossip Window and Gossip texts, Creature Template reload command, Powers display, Group invites, Phaseshifting, Dungeon Encounters sorting, Override Spells (Mechanic Abilities), Stealth flags and some other stuff.

+++ src/server/game/Handlers/PetHandler.cpp
@@ -1052,7 +1052,7 @@ void WorldSession::HandlePetCastSpellOpcode(WorldPacket& recvPacket)
         
         recvPacket.ReadByteSeq(moveGuid1[4]);
         
-        for (int i = 0; i < unkCount; ++i)
+        for (uint8 i = 0; i < unkCount; ++i)
         {
             recvPacket.read_skip<uint32>();
         }
